# Fully port CheckMap to Rust