--- a/qiskit/__init__.py
+++ b/qiskit/__init__.py
@@ -87,6 +87,7 @@ sys.modules["qiskit._accelerate.synthesis.permutation"] = _accelerate.synthesis.
 sys.modules["qiskit._accelerate.synthesis.linear"] = _accelerate.synthesis.linear
 sys.modules["qiskit._accelerate.synthesis.clifford"] = _accelerate.synthesis.clifford
 sys.modules["qiskit._accelerate.synthesis.linear_phase"] = _accelerate.synthesis.linear_phase
+sys.modules["qiskit._accelerate.check_map"] = _accelerate.check_map
 
 from qiskit.exceptions import QiskitError, MissingOptionalLibraryError